Flyweight Pattern
Turned all the observers into a child of a observer flyweight so there is less repeated code and much easier to edit any observer
