+++ Assets/Movement/WallRunning.cs
@@ -3,20 +3,8 @@ using System.Collections.Generic;
 using System.Security.Cryptography;
 using UnityEngine;
 
-public class WallRunning : MonoBehaviour, IObserver
+public class WallRunning : ObserverFlyweight
 {
-	[Header("Observer")]
-	public Win[] Observees;
-
-	public void UpdateWhenNotified()
-	{
-		foreach (var observee in Observees)
-		{
-			observee.Unsubscribe(this);
-		}
-		this.enabled = false;
-	}
-
 	[Header("WallRunning")]
 	public LayerMask isWall;
 	public LayerMask isGround;
@@ -77,16 +65,11 @@ public class WallRunning : MonoBehaviour, IObserver
 		}
 	}
 
-	private void Start()
+	protected override void Start()
 	{
 		rb = GetComponent<Rigidbody>();
 		pm = GetComponent<PlayerMovement>();
 		lg = GetComponent<LedgeGrabbing>();
-
-		foreach (var observee in Observees)
-		{
-			observee.Subscribe(this);
-		}
 	}
 
 	private void Update()
