+++ Assets/Scripts/ObserveeFlyweight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 58538e10502288a458b9b7f2c754b0d2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
