+++ Assets/Scripts/ObserverFlyweight.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 80bfe9d576ab4b24289138de4f2e1ad0
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
